Added Floyds Cycle Finding Algorithm
Floyd's cycle-finding algorithm is a pointer algorithm that uses only two pointers, which move through the sequence at different speeds. It is also called the "tortoise and the hare algorithm", alluding to Aesop's fable of The Tortoise and the Hare.  It is a simple implementation O(n) algorithm in C to determine a cycle using linked list data structure.